changed how we call certain reservation functions -- makes more sense this way
